Fixxed more warnings (IDE0044 IDE0038  IDE0031 IDE0039 IDE0019 IDE0004 IDE0059 IDE0062 IDE0065)

+++ ClosedXML/Excel/CalcEngine/CalcEngineHelpers.cs
@@ -138,7 +138,7 @@ namespace ClosedXML.Excel.CalcEngine
             var range = (rangeExpression?.Value as CellRangeReference)?.Range;
             if (range == null)
                 return 0;
-            return (long)(range.LastColumn().ColumnNumber() - range.FirstColumn().ColumnNumber() + 1) *
+            return (range.LastColumn().ColumnNumber() - range.FirstColumn().ColumnNumber() + 1) *
                    (long)(range.LastRow().RowNumber() - range.FirstRow().RowNumber() + 1);
         }
     }
